Reworked Settings.ini handling
Will now create a fresh copy of the settings.ini with fresh install defaults.

Tightened the title match for closing GottaGoFast.ahk and will only close the script itself (Not editors with its name in it XD)

Fixed a lot of formatting issues with the code. several areas had incorrect notation.