Bugfix and Added variable for detonate mines Issues saving hotkeys is resolved, if you had the issue of it saving to your INI file as ERROR please delete/edit your settings file to fix it.
Also if you are having issues with detonate mines. I have added a
DetonateHex variable with the colorhex it is searching for. Edit this
color if you are not having it detonate. to find the color go to
1542, 901 or 1658, 901 location and copy the hex code from window spy
BACKWARDS in chunks of two digits.. (not sure why)
372041 in window spy becomes 0x412037 in the script

I will add a way to sample the detonate mines color eventually when I
do a rework of the failsafe color grab screens. But for now will need
to be done manually.