Rewrote Failsafe and several tuning adjustments Improved the usability of the "failsafe" section and reworded the section to improve the understanding of what it does. It now allows the user to recallibrate any of the colors individually. I will do the same thing for detonate mines.
Swapped the order of the colors in LootColors array, making black the
first to search for improves the pickup since most items are black BG.

Added the hints from version 1.2 of wingman into the GUI to help people
set up their keybinds properly.

Added enable/disable option for keys Alt(1-10) for stash tabs

Added check for last second before sending click in the LootScan,
hopefully this will improve pickup when tapping the Item Pickup key
while cursor is on item.

Changed the direction of the ItemSort function, it now goes vertically
then horizontally, was unable to do this before because of tooltip
getting in the way. Just created a random area to place the mouse when
it finished a column. This will speed the scan up drastically when
dealing with large items taking up inventory. 20-30% faster

Padded latency between click down and click up in all functions by 15.
This is to make sure small latency discrepancy wont affect script.