Added profile system There are now three buttons to quickly load profiles 1-3 and another tab to configure up to 10 profiles for different characters.
On the Profile management tab you can edit the names of the profiles to
something more memorable. Simply edit the values in the box and it will
automatically save any changes made inside.

To use the quick load buttons you need to first save to the profile in
the management tab. If there is no profile existing when attempting to
load one, it will exit from loading error values.

For anyone who wants to poke around in the settings file, the profile
variables are saved into their own sections: profile1, profile2 etc