Impliment aspect ratio dropdown box Added a dropdown box which can easily add any number of alternate aspect ratios. Standard aspect ratio is based on 1920x1080, UltraWide is 3840x1080, and QHD will be 2650x1440 when it is fully implimented.
I will be looking for someone to volunteer for input of the 4:3 ratios
like 800x600 or other similar resolution.

Bugfix for randomly dissapearing Profile names on reload.
The issue was that they were saving the entire GUI when you edited the
any of the boxes. This was causing randomly blank values to appear
instead of being read from the ini. They now only interact w/ themself.

More finetuning of the code, adding missing returns to 4 locations