Instant Flasks on attack keys and assorted fixes Complete rework in how the attack flasks were being handled. Instead of having them being included inside of the pixel check timer, I have moved them into the hotkey group. They now directly fire a function which will fire its assigned flasks. The function makes sure you are not in chat, stash, hideout, vendor or not on character.
tweeked the loot vacuum with some sleep, -1 in specific lines after
checking for the key being held down. This has improved the accuracy
of the vacuum when tapping the key or letting go at odd times.

Adjustments to he profile loading and saving to match the changes to
the attack keys. Old profiles will load RButton and W as default to
avoid loading an ERROR value because of missing items. Newly saved
profiles will not have any issue.