Changes to menu behavior Now that I have ironed out the kinks in checking for the POE executable, I have gone ahead and removed the necessity of restarting the script to save your configuration. If the user decides to press the save button when there is no game running, it will warn them that they will not resample the life colors but that all other settings will be saved. When the game is open it will inform then they have just resampled those colors and they should have been on a character.
I have done the same thing for the calibration buttons, they will not
work if the game is not running, and no longer need to have the game
open right when the script starts.

These changes will help make the interface a bit more user friendly,
it was annoying to keep needing to reload the script just to save.

Added an ability for the master script to have the subscripts reload
their settings instead of restarting them. Now you no longer need to
re-enable autoquicksilver if you had it turned on when you save.