Cooldowns from main script will go directly to sub Now that I have a method of interaction with subscript, I have removed duplicate functions from the subscript and replaced them with similar code that runs on receiving a message instead.
Esentially, it was putting the flasks on two slightly different timers
sometimes. Now it will go 1:1 with each other. Except when the subscript
is triggering the CD, I couldnt get it to work going from the subscript
back to the main. That will be something to possibly figure out later.

Other things to note:
I found a way to improve the Quicksilver flasks, they are not being put
into a list of available cooldowns, and he was using a long sleep
instead of a timer. I think removing the sleep and making it fill a list
of ready to pop flasks. then another section will only run if there are
items in that array, and the other portion will stop filling the list.
That way it can keep checking for the next flask to pop and you holding
the key, and consistantly moving from one flask to the next without
skipping them. Will look at syntax of adding and removing into array.