Created a new utility Function It can now be expanded to support any number of utilities, input name of the Utility to fire and it manages all the cooldowns and keys. I put all the settings together for easily adding more.
As a proof of concept I have added Phase run into a new Utility tab.
Choose the key to fire, select checkbox to enable and input cooldown.
It is currently configured to only fire when popping a quicksilver, will
expand on those options later.

Now the next step is choosing which utility buffs would be useful for
autocast! Next in mind is to add in Vaal Discipline when low on ES.