Massive new data files I spent a lot of time this weekend scraping the wiki's to gather data for the script. What I ended up doing was learning an entirely new lib for scraping websites and getting pretty efficient at it! :)
With this new data we can finally figure out which is a Suffix or Prefix
and also how many open affix slots there are (with known 2x/3x/4x mods)

As of right now this is the data dump of all the equipment and every
possible affix that they can roll. I still need to do flasks and jewels.