I added a "Main Attack Only" Checkmark in the utility menu, this is so triggers dont happen all the time, but only while main attacking. This differs from main, because main casts all the time during main. You can not for instance trigger a Text Search, but only when main attacking. Use full for buffs/totems/hexes/etc...
I added an "Unquality flask" option to Inventory-> Stash Tabs. This is for leveling so that you can deposit all your flasks and upgrade them end of act or whenever.

MainAttackOnly is by default checked and sometimes needs to be unchecked and rechecked. I need to find how to properly save this (saveload?, ini edit?) Would appreciate the help.